Finish implementing `mention_text`.
Beyond implementing `mention_text`, I have also added a test case
for it with a mocking implementation of a minimal puppet. I hope
that we can build upon this approach to improve test coverage.

Note that the current implementation is kind of cumbersome and
it's coupled with the message module specifically. We may need
to make it workable when testing other modules.

Fixes #178